dtype issues for replicate padding

+++ code/Evaluation.py
@@ -114,7 +114,8 @@ with torch.no_grad():  # save the images
     if input_with_noise:
         input_size = im_3d.size()
         # make noise channel and concatenate it to input:
-        noise = torch.randn(input_size[0], 1, *input_size[2:], device=device)
+        noise = torch.randn(input_size[0], 1, *input_size[2:],
+                            device=device, dtype=im_3d.dtype)
         im_3d = torch.cat((im_3d, noise), dim=1)
 
     # nz1, nz2, nz3 = im_3d.size()[-3:]
